✨ force external links to be in a separate block
