+++ db/model/Gdoc/exampleEnrichedBlocks.ts
@@ -297,12 +297,6 @@ export const enrichedBlockExamples: Record<
                 subtitle: "This is an inspiring chart",
                 type: "recirc-link",
             },
-            {
-                url: "https://example.com/",
-                title: "Example",
-                subtitle: "This is an external link",
-                type: "recirc-link",
-            },
         ],
         parseErrors: [],
     },
